+++ Elements/src/Spatial/Grid2d.cs
@@ -727,8 +727,9 @@ namespace Elements.Spatial
         /// <summary>
         /// Test if the cell is trimmed by a boundary.
         /// </summary>
+        /// <param name="treatFullyOutsideAsTrimmed">Should cells that fall entirely outside of the boundary be treated as trimmed? True by default.</param>
         /// <returns>True if the cell is trimmed by the grid boundary.</returns>
-        public bool IsTrimmed()
+        public bool IsTrimmed(bool treatFullyOutsideAsTrimmed = true)
         {
             if (boundariesInGridSpace == null || boundariesInGridSpace.Count == 0)
             {
@@ -736,10 +737,28 @@ namespace Elements.Spatial
             }
 
             var baseRect = GetBaseRectangleTransformed();
+            if (treatFullyOutsideAsTrimmed && IsOutside(baseRect))
+            {
+                return true;
+            }
             var trimmedRect = Polygon.Intersection(new[] { baseRect }, boundariesInGridSpace);
             if (trimmedRect == null || trimmedRect.Count < 1) { return false; }
             if (trimmedRect.Count > 1) { return true; }
             return !trimmedRect[0].IsAlmostEqualTo(baseRect, Vector3.EPSILON);
+        }
+
+        /// <summary>
+        /// Test if the cell is fully outside the boundary.
+        /// </summary>
+        /// <returns>True if the grid cell is totally outside the boundary.</returns>
+        public bool IsOutside()
+        {
+            if (boundariesInGridSpace == null || boundariesInGridSpace.Count == 0)
+            {
+                return false;
+            }
+            var baseRect = GetBaseRectangleTransformed();
+            return IsOutside(baseRect);
         }
 
         #endregion
@@ -757,6 +776,22 @@ namespace Elements.Spatial
                 throw new NotSupportedException("An attempt was made to modify the underlying U or V grid of a 2D Grid, after some of its cells had already been further subdivided.");
             }
             this.cells = null;
+        }
+
+        private bool IsOutside(Polygon baseRect)
+        {
+            var perimeter = boundariesInGridSpace.First();
+            // if any vertex of the rect is fully outside, we are trimmed
+            perimeter.Contains(baseRect.Vertices.First(), out var containment);
+            if (containment == Containment.Outside)
+            {
+                return true;
+            }
+            if (boundariesInGridSpace.Count() > 1)
+            {
+                return boundariesInGridSpace.Skip(1).Any(boundary => boundary.Covers(baseRect));
+            }
+            return false;
         }
 
         private void InitializeUV(Domain1d uDomain, Domain1d vDomain)
@@ -915,6 +950,13 @@ namespace Elements.Spatial
         }
 
         /// <summary>
+        /// The default extension tolerances of Vector3.EPSILON were creating funny conditions
+        /// when gridlines were close to but not quite parallel to a polygon edge.
+        /// We pass along a much smaller tolerance when we run our line extensions for Grid2d.
+        /// </summary>
+        private const double ExtensionTolerance = Vector3.EPSILON * Vector3.EPSILON;
+
+        /// <summary>
         /// Modifies a list of lines intended to represent uv guides in place to hit the bounds.
         /// Accounts for skewed, parallel lists of 2. If list contains more lines, those will be ignored.
         /// </summary>
@@ -927,7 +969,7 @@ namespace Elements.Spatial
 
             for (var i = 0; i < lines.Count(); i++)
             {
-                lines[i] = lines[i].ExtendTo(boundary, true, true);
+                lines[i] = lines[i].ExtendTo(boundary, true, true, ExtensionTolerance);
             }
 
             var new1 = ExtendLineSkewed(bounds, lines[0], lines[1]);
@@ -962,12 +1004,12 @@ namespace Elements.Spatial
                 // move to start and extend
                 var toStart = possiblySkewedLine.Start - intersection;
                 newLine = newLine.TransformedLine(new Transform(toStart));
-                newLine = newLine.ExtendTo(boundary, true, true);
+                newLine = newLine.ExtendTo(boundary, true, true, ExtensionTolerance);
 
                 // move to end and extend
                 var toEnd = possiblySkewedLine.End - possiblySkewedLine.Start;
                 newLine = newLine.TransformedLine(new Transform(toEnd));
-                newLine = newLine.ExtendTo(boundary, true, true);
+                newLine = newLine.ExtendTo(boundary, true, true, ExtensionTolerance);
 
                 // move back to original
                 var toBeginning = intersection - possiblySkewedLine.End;