+++ CHANGELOG.md
@@ -1,23 +1,53 @@
 # Changelog
 
+## 0.9.0
+
+### Added
+
+- `Grid2d.IsOutside()`
+
+### Removed
+- `BuiltInMaterials.Dirt`
+- `BuiltInMaterials.Grass`
+
+### Changed
+
+- `Grid2d.IsTrimmed()` now takes an optional boolean parameter `treatFullyOutsideAsTrimmed`
+- `ConstructedSolid` serializes and deserializes correctly.
+- `Solid.AddFace(Polygon, Polygon[])` can take an optional third `mergeVerticesAndEdges` argument which will automatically reuse existing edges + vertices in the solid.
+- Adds optional `tolerance` parameter to `Line.ExtendTo(Polygon)`, `Line.ExtendTo(IEnumerable<Line>)`, `Vector3.IsParallelTo(Vector3)`.
+
+### Fixed
+
+- Fixed a bug in `ConvexHull.FromPoints` when multiple X coordinates are equal.
+- Fixed a bug in `Grid2d(Polygon, Vector3, Vector3, Vector3)` where U or V directions skew slightly when they nearly parallel with a boundary edge.
+
+
 ## 0.8.5
 
 ### Added
+
 - `Elements.Spatial.CellComplex`
 - `Grid2d.ToModelCurves()`
 - Alpha release of `Hypar.Elements.Components`
 - `Polyline.OffsetOnSide`
 - `Ray.Intersects(Polygon)`
 - `Vector3.DistanceTo(Polygon)`
+- `(double Cut, double Fill) Topography.CutAndFill(Polygon perimeter, double topElevation, out Mesh cutVolume, out Mesh fillVolume, double batterAngle)`
+- `void Topography.Trim(Polygon perimeter)`
+
 ### Changed
+
 - `ColorScale` no longer bands colors but returns smooth gradient interpolation. It additionally now supports a list of values that correspond with the provided colors, allowing intentionally skewed interpolation.
 - `Solids.Import` is now public.
 - `Polygon.Contains` was modified to better handle polygons that are not on the XY plane.
 
 ### Fixed
+
 ## 0.8.4
 
 ### Added
+
 - `BBox3.IsValid()`
 - `BBox3.IsDegenerate()`
 - `Elements.Light`
@@ -41,6 +71,7 @@
 - Release helper github action
 
 ### Changed
+
 - `Elements.DirectionalLight` now inherits from `Elements.Light`.
 - `Elements.ContentCatalog` now has a `ReferenceConfiguration` property.
 - `SHSProfile`
@@ -59,6 +90,7 @@
 ## 0.8.3
 
 ### Added
+
 - `Profile.ToModelCurves()`
 - `Profile.Difference()`
 - `Profile.Intersection()`